Split graphicsviewlib files
