+++ scripts/addie
@@ -89,14 +89,14 @@ import addie.utilities.specify_plots_style as ps
 import addie.calculate_gr.edit_sq_dialog
 
 from addie.calculate_gr.pdf_lines_manager import PDFPlotManager
-from addie.graphicviewlib import GofRView, SofQView
 from mantidqt.widgets.jupyterconsole import InProcessJupyterConsole
 
 from addie.rietveld.braggview import BraggView
-
 from addie.rietveld.braggtree import BraggTree
-from addie.calculate_gr.gofrtree import GofRTree
 
+from addie.calculate_gr.gofrtree import GofRTree
+from addie.calculate_gr.gofrview import GofRView
+from addie.calculate_gr.sofqview import SofQView
 
 from addie.icons import icons_rc
